refactor: overhaul source plane coordinate handling
This commit overhauls source plane coordinate handling across the codebase:
- add new `infer_source_bbox` utility to compute padded physical bounds from ray-traced points
- update `build_source_grid` to accept explicit (xmin, xmax, ymin, ymax) bounds instead of half-size
- refactor simulator, prob model, and source model classes to use bounding boxes instead of half-size
- update regularization builders to use physical grid bounds for accurate scaling
- fix linear solver to cast output arrays to the specified input dtype
- update all downstream code including demos, tests, and pipeline workflows

+++ TinyLensGpu/utils/linear_solver.py
@@ -69,7 +69,7 @@ def fnnls_jax(Z: Array, x: Array) -> Tuple[Array, float]:
         d_unscaled = d_scaled / safe_norms
         d = jnp.where(active, d_unscaled, 0.0)
         res = jnp.linalg.norm(x - Z @ d)
-        return d, res
+        return d.astype(dtype), res.astype(dtype)
 
     should_solve = jnp.any(active) & jnp.any(jnp.abs(ZTx) > 0)
     return jax.lax.cond(should_solve, solve_active_system, zero_solution, operand=None)
